% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneBoostTreeBayesian.R
\name{TuneBoostTreeBayesianUltra}
\alias{TuneBoostTreeBayesianUltra}
\title{Executar tuning ultraotimizado de boosted trees}
\usage{TuneBoostTreeBayesianUltra(formula, data, initial = 20L, nIter = 60L, engine = "lightgbm", command = NULL, strict_limbo = TRUE)}
\arguments{\item{formula, data, initial, nIter, engine}{Argumentos repassados a \code{TuneBoostTree()}.}\item{command, strict_limbo}{Argumentos repassados a \code{TuneBoostTreeUltraConfig()}.}}
\value{Objeto \code{tbtb_tune_result}, mesmo retorno de \code{TuneBoostTree()}.}
\description{Atalho que monta \code{TuneBoostTreeUltraConfig()} e executa \code{TuneBoostTree()}.}
